Remove version from composer.json
According to the `composer.json` schema, the version should be omitted from the `composer.json` file whenever it can be inferred from somewhere. We use annotated tags for each new release of php-cypher-dsl, so Composer will now infer the version from there. This reduces the risk of problems resulting from forgetting to update the tag.